task_2/requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.27.0
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
colorama==0.4.6
factory-boy==3.2.1
Faker==15.1.4
fastapi==0.86.0
greenlet==2.0.0.post0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
mock==4.0.3
packaging==21.3
pluggy==1.0.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.1
python-dateutil==2.8.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
